refactor(workflow): move workflow YAML schema into workflow module
`library/` owns the file abstraction (`UpstreamOp`, `SyncedFile`,
`ParsedFile`); per-entity wire formats belong with their entity.

Moves `WorkflowYaml` / `WorkflowSandboxYaml` (with `ScratchYamlConfig`,
`RepoYamlConfig`) / `WorkflowTriggerYaml` / `WorkflowStepYaml`,
`render_workflow_yaml`, `parse_workflow_yaml`,
`workspace_name_from_workflow_path`, `ParsedWorkflowFile`, and the
test-helper `upstream_op_for_workflow` from `library/file.rs` into a
new `core/src/workflow/yaml.rs`. Tests move with them.

Library exposes `name_from_filename` and `slugify` as `pub(crate)` so
the workflow module (and any future entity yaml module) can share
them. Drops the workflow re-exports from `library/mod.rs`.

Production callers (`WorkflowDefinition::LibrarySynced::render`,
`WorkflowLibraryImporter::parse` / `upsert_in_op`) now go through
`super::yaml::*` instead of `crate::library::*`.
